### 2025-09-25 #309 @editorialbot
[REVIEW]: nwb4edu: an Online Textbook for Teaching and Learning with NWB Datasets
**Submitting author:** @ajuavinett (<a href="http://orcid.org/0000-0002-4254-3009">Ashley Juavinett</a>)
**Repository:** https://github.com/nwb4edu/nwb4edu.github.io
**Branch with paper.md** (empty if default branch): master
**Version:** 1.1.0
