### 2025-09-25 nwb4edu/nwb4edu.github.io#3 @arokem
[JOSE Review] Documentation
Hello! This looks interesting and potentially very useful. The rendered website is beautiful with tons of great content (but lookout for some future issues about some gremlins), but the repository provides no "meta-documentation" to help orient potential users to the content, how it gets put together into the website, and how someone might reuse the content and/or contribute to it. I think that some more content in the README might be appropriate. Also, a link to the rendered book website on the repo front page might be helpful.
