### 2025-09-25 nwb4edu/nwb4edu.github.io#4 @arokem
[JOSE Review] A few book rendering gremlins
There are some (unintentional?) errors on a few pages of the rendered book:

- https://nwb4edu.github.io/Lesson_1/02-Working_with_NWB_format_in_Python.html (file not found error in the second code cell).
- https://nwb4edu.github.io/Lesson_2/01-SpikeSorting.html (missing data in all figures).
- https://nwb4edu.github.io/Lesson_2/02-Visualizing_Recordings.html (name error in cell 2)
- https://nwb4edu.github.io/Lesson_3/02-Analyzing_Computed_Features.html (keyboard interrupt in cell 2 - probably a time-out)
- https://nwb4edu.github.io/Lesson_4/01-Retrieving_Brain_Observatory_Data.html (index error in last cell)
- https://nwb4edu.github.io/Lesson_5/01-PlaceFields.html (keyboard interrupt in last cell)
- https://nwb4edu.github.io/Lesson_5/02-RotationalDynamics.html (keyboard interrupt in cell 4)
