### 2025-09-25 nwb4edu/nwb4edu.github.io#5 @arokem
[JOSE Review] License
I believe that a creative commons license may be more appropriate for most of the content of this repository (rather than the currently used MIT license), but nbd.
