### 2025-09-26 nwb4edu/nwb4edu.github.io#5 @arokem
Also, the paper claims the material is released under CC-BY-SA, but the repo has the MIT license in it, so maybe I am missing something.
