### 2025-09-26 nwb4edu/nwb4edu.github.io#6 @arokem
[JOSE Review] Dependencies
> Is there a clearly stated list of dependencies?

While the rendered website contains a list of dependencies to run the code, the repo does not have a list of dependencies to build the rendered website. I think that documentation of this (even brief) should be included here.
