### 2025-09-26 nwb4edu/nwb4edu.github.io#7 @arokem
[JOSE Review] Community guidelines
Per review checklist:

> Are there clear guidelines for third parties wishing to 1) Contribute to the module 2) Report issues or problems with the module 3) Seek support

I couldn't find this.
