### 2025-09-26 #309 @arokem
OK - I am done with my initial review. The materials are really great, and will be useful to many instructors. You can find a few issues and suggestions here:

- [x] https://github.com/nwb4edu/nwb4edu.github.io/issues/3
- [x] https://github.com/nwb4edu/nwb4edu.github.io/issues/4
- [x] https://github.com/nwb4edu/nwb4edu.github.io/issues/5
- [x] https://github.com/nwb4edu/nwb4edu.github.io/issues/6
- [x] https://github.com/nwb4edu/nwb4edu.github.io/issues/7

Once these are addressed, I think that I can probably check off the other boxes as well.
