### 2025-10-05 nwb4edu/nwb4edu.github.io#4 @TomDonoghue
Also from the JOSE review, some additional code block quirks I noticed:

- Setup page: last code block prints an error (I think unintentionally)
- Setup page, code block typo: `!pip install pwynb`
- In Lesson_1/02, many of the code blocks have no output after the error mentioned above, so these need re-executing & checking
