### 2025-10-05 #309 @TomDonoghue
I have also finished my initial review of this resource. I agree with Ariel's comments & topics raised and added a couple comments to those existing issues, and added a couple more. Broadly speaking, I think this is a great resource, and very good from a student facing perspective, but could do with a few updates to make the resource a bit more sign-posted from the teaching / developer side of things. Related to these comments, I've left a few checkboxes above open for now - I think they can all be addressed with relatively minor updates to the resource.
