+++ README.md
@@ -1,0 +1,212 @@
+# nwb4edu: Teaching and Learning with NWB Datasets
+
+[![License: CC BY-SA 4.0](https://img.shields.io/badge/License-CC%20BY--SA%204.0-lightgrey.svg)](https://creativecommons.org/licenses/by-sa/4.0/)
+
+An interactive online textbook for teaching and learning neuroscience using Neurodata Without Borders (NWB) datasets.
+
+**Live Website:** [https://nwb4edu.github.io](https://nwb4edu.github.io)
+
+## About This Project
+
+The `nwb4edu` online textbook provides interactive and self-guided resources for educators, students, and self-guided learners who would like to use open neuroscience datasets for research and/or teaching. The textbook introduces users to Neurodata Without Borders (NWB), a data standard for neurophysiology data, and teaches how to access, analyze, and visualize NWB datasets using Python.
+
+### Key Features
+
+- **Interactive Jupyter notebooks** with executable code
+- **Comprehensive tutorials** covering data access, analysis, and visualization
+- **Problem sets** for hands-on practice
+- **Real datasets** from DANDI Archive and Allen Institute
+- **Greatest Hits lessons** that recreate figures from landmark neuroscience papers
+- **Launch on JupyterHub** - no local installation required
+
+## Content Overview
+
+The textbook is organized into several sections:
+
+### Introduction
+- Using this book
+- Environment setup
+- Resources for educators
+
+### Data Science in Python (Refresher)
+- NumPy
+- Pandas
+- Matplotlib
+- SciPy
+
+### PyNWB Lessons
+**Lesson 1:** Introduction to NWB format
+- Obtaining datasets with DANDI
+- Working with NWB format in Python
+- Interactive NWB data exploration
+
+**Lesson 2:** Large-scale electrophysiology
+- Spike sorting
+- Visualizing extracellular recordings
+
+### AllenSDK Lessons
+**Lesson 3:** Single-cell electrophysiology
+- Obtaining data for single cells
+- Analyzing computed features
+
+**Lesson 4:** Two-photon calcium imaging
+- Retrieving Brain Observatory data
+- Analyzing two-photon data
+- Correlations in neural activity
+
+### Greatest Hits in Neuroscience
+Recreate figures from foundational papers:
+- Place fields (O'Keefe & Dostrovsky, 1971; Grosmark & Buzsáki, 2016)
+- Rotational dynamics in motor cortex (Churchland et al., 2012)
+
+## Using This Textbook
+
+### For Students
+
+The easiest way to use this textbook is through the live website at [https://nwb4edu.github.io](https://nwb4edu.github.io). You can launch any notebook directly on JupyterHub using the rocket icon in the top right corner of each page.
+
+No installation is required when using JupyterHub!
+
+### For Local Development
+
+If you want to run the notebooks locally:
+
+1. Clone this repository:
+```bash
+git clone https://github.com/nwb4edu/nwb4edu.github.io.git
+cd nwb4edu.github.io
+```
+
+2. Install dependencies:
+```bash
+pip install -r requirements.txt
+```
+
+3. Launch Jupyter:
+```bash
+jupyter notebook
+```
+
+### For Educators
+
+This textbook is designed to be modular - you can use individual lessons or entire sections in your courses. All materials are licensed under CC-BY-SA-4.0, so you're free to adapt and modify them for your teaching needs.
+
+See our [For Educators](https://nwb4edu.github.io/Introduction/For_Educators.html) guide for more information on integrating these materials into your curriculum.
+
+## Building the Book
+
+This textbook is built using [Jupyter Book](https://jupyterbook.org/).
+
+### Requirements
[…]
