+++ requirements.txt
@@ -1,0 +1,31 @@
+# Core dependencies for running nwb4edu notebooks
+
+# NWB and data access
+pynwb>=2.0.0
+dandi>=0.45.1
+h5py>=3.0.0
+hdmf>=3.0.0
+
+# Allen Institute SDK
+allensdk>=2.11.2
+
+# Scientific computing
+numpy>=1.18.5,<1.24.0
+pandas>=1.0.0
+scipy>=1.6.0
+xarray>=0.15.1
+
+# Visualization
+matplotlib>=3.4.0
+seaborn>=0.11.0
+
+# Widgets for interactive exploration
+nwbwidgets>=0.8.0
+ipywidgets>=7.6.0
+
+# Spike analysis
+spykes>=0.3.dev0
+
+# Jupyter
+jupyter>=1.0.0
+notebook>=6.0.0
