### 2025-10-05 nwb4edu/nwb4edu.github.io#3 @ajuavinett
Thanks, @arokem! Agreed about the needed meta-documentation. To address this I've done the following:

1. Created a [README.md](https://github.com/nwb4edu/nwb4edu.github.io/tree/master)
2. Added [CONTRIBUTING.md](https://github.com/nwb4edu/nwb4edu.github.io/blob/master/CONTRIBUTING.md) (also in response to https://github.com/nwb4edu/nwb4edu.github.io/issues/7)
3. Added dependency documentation:
- requirements.txt - all code execution dependencies
- requirements-docs.txt - dependencies for building the documentation

The README now prominently links to the rendered book at https://nwb4edu.github.io and provides guidance for all three audiences: students using the materials, educators adapting them and/or contributing to the project.
