### 2025-10-05 nwb4edu/nwb4edu.github.io#5 @ajuavinett
Oh good catch, oops. License changed to CC-BY-SA!
