### 2025-10-05 2498d83
Simplify Setup.ipynb package dependencies
- Reduced package list to essential top-level dependencies only
- dandi, allensdk, nwbwidgets, seaborn automatically install subdependencies
- Updated code cells to check for essential packages
- Clarified which dependencies are included with each package
- Added Lesson 5 packages: pynapple, plotly, fsspec

🤖 Generated with [Claude Code](https://claude.com/claude-code)

Co-Authored-By: Claude <noreply@anthropic.com>
