+++ _config.yml
@@ -4,7 +4,12 @@ only_build_toc_files: true
 repository:
   url: https://github.com/nwb4edu/nwb4edu.github.io
   branch: master
-launch_buttons: 
+launch_buttons:
   colab_url: "https://colab.research.google.com"
   thebe                  : true
   jupyterhub_url: "https://hub.dandiarchive.org/"
+
+execute:
+  execute_notebooks: cache
+  allow_errors: true
+  timeout: 180
