### 2025-10-05 be204cc
fixed lesson 1
