### 2025-10-06 nwb4edu/nwb4edu.github.io#4 @ajuavinett
Thanks for your careful read both -- very embarrassing these errors were still here!

- I fixed issue with file missing in **Working with NWB format** to provide an error message if user did not complete notebook 01 (and get file) as instructed. Similar fix for the **Visualizing Recordings** page.
- Updated the kind of data streaming because ros3 is no longer the recommended one, nor does it work in dandihub. This fixed issues in the spike sorting notebook. Hoping remfile is fairly stable but might also consider adding fsspec option.

(to be continued, working on other fixes...)
