### 2025-10-06 f678734
Update requirements.txt to match Setup.ipynb changes
- Simplified to list only top-level dependencies
- Documented which packages are included as subdependencies
- Added remfile for data streaming
- Added Lesson 5 packages: pynapple, plotly, fsspec
- Removed redundant subdependencies (pynwb, h5py, hdmf, numpy, pandas, scipy, matplotlib, ipywidgets)
- Matches the dependency structure in Setup.ipynb

🤖 Generated with [Claude Code](https://claude.com/claude-code)

Co-Authored-By: Claude <noreply@anthropic.com>
