### 2025-10-06 nwb4edu/nwb4edu.github.io#6 @ajuavinett
Thanks for this -- wrangling the dependencies for all of these different notebooks has been tough but I think I was able to succinctly summarize it on the [Setup](https://nwb4edu.github.io/Introduction/Setup.html) page, which just focuses on the top-level packages. These changes are also reflected in the requirements.txt file.

I've also added plotly installation to Lesson_5/01.
