### 2025-10-06 8fd454a
Update paper.md
add information about dataset types
