### 2025-10-06 #309 @ajuavinett
Agreed, thank you both for these really detailed reviews. This is my first big step into more formally sharing an open source coding project and your reviews have been wonderfully instructive in my own development. I'm working through the issues you've raised and commenting on those as I do.
