### 2025-10-06 9b151c4
add jupyterbook ref and info about datasets
