+++ paper.md
@@ -23,7 +23,7 @@ bibliography: paper.bib
 ---
 
 # Summary
-The ``nwb4edu`` online textbook provides an interactive set of lessons for educators, students, and self-guided learners who would like to use open neuroscience datasets for research and/or teaching. Specifically, it introduces users to Neurodata Without Borders (NWB), a relatively new data format for neurophysiology data [@rubel_neurodata_2022; @teeters_neurodata_2015] which can be accessed via [DANDI (Distributed Archives for Neurophysiology Data Integration)](https://about.dandiarchive.org/). The ``nwb4edu`` textbook contains tutorials in the form of interactive Python notebooks that enable users to wrangle, analyze, and visualize NWB datasets which contain behavioral, electrophysiological, and/or imaging time series. Importantly, it also provides users with the necessary conceptual background to understand how the data was collected and how it can be analyzed. The broad understanding of how a dataset is structured and how to navigate a dataset can be applied to other data standards, while the in-depth understanding of NWB datasets can empower students and researchers to conduct their own research using these data. 
+The ``nwb4edu`` online textbook provides an interactive set of lessons for educators, students, and self-guided learners who would like to use open neuroscience datasets for research and/or teaching. Specifically, it introduces users to Neurodata Without Borders (NWB), a relatively new data format for neurophysiology data [@rubel_neurodata_2022; @teeters_neurodata_2015] which can be accessed via [DANDI (Distributed Archives for Neurophysiology Data Integration)](https://about.dandiarchive.org/). The ``nwb4edu`` textbook contains tutorials within an interactive JupyterBook [@executable_books_project_2021_4539666] that enable users to wrangle, analyze, and visualize NWB datasets. The datasets featured in this resource currently include rodent behavioral data, single-cell and extracellular electrophysiological recordings, and two-photon imaging experiments. Importantly, it also provides users with the necessary conceptual background to understand how the data was collected and how it can be analyzed. The broad understanding of how a dataset is structured and how to navigate a dataset can be applied to other data standards, while the in-depth understanding of NWB datasets can empower students and researchers to conduct their own research using these data. 
 
 Each lesson contains learning objectives, tasks, and a problem set so that students can test and extend their skills. Executable notebooks can be launched on a pre-configured JupyterHub, making these easily accessible to anyone with an internet connection. All of the materials are covered by a CC-BY-SA-4.0 and are available at http://nwb4edu.github.io. 
 
