+++ paper.bib
@@ -36,3 +36,14 @@
 	pages = {629--634},
 	file = {ScienceDirect Full Text PDF:/Users/ashley/Zotero/storage/5U2E4EAK/Teeters et al. - 2015 - Neurodata Without Borders Creating a Common Data .pdf:application/pdf;ScienceDirect Snapshot:/Users/ashley/Zotero/storage/54QJCMZS/S0896627315009198.html:text/html},
 }
+
+@software{executable_books_project_2021_4539666,
+  author       = {Executable Books Project},
+  title        = {Jupyter Book},
+  month        = feb,
+  year         = 2021,
+  publisher    = {Zenodo},
+  version      = {v0.10.0},
+  doi          = {10.5281/zenodo.4539666},
+  url          = {https://doi.org/10.5281/zenodo.4539666}
+}
