### 2025-10-06 671fdaf
add dataset references
