### 2025-10-06 41ab0ec
Update paper.bib
