+++ paper.bib
@@ -47,3 +47,132 @@
   doi          = {10.5281/zenodo.4539666},
   url          = {https://doi.org/10.5281/zenodo.4539666}
 }
+
+
+@article{de_vries_sharing_2023,
+	title = {Sharing neurophysiology data from the {Allen} {Brain} {Observatory}},
+	volume = {12},
+	issn = {2050-084X},
+	url = {https://doi.org/10.7554/eLife.85550},
+	doi = {10.7554/eLife.85550},
+	abstract = {Nullius in verba (‘trust no one’), chosen as the motto of the Royal Society in 1660, implies that independently verifiable observations—rather than authoritative claims—are a defining feature of empirical science. As the complexity of modern scientific instrumentation has made exact replications prohibitive, sharing data is now essential for ensuring the trustworthiness of one’s findings. While embraced in spirit by many, in practice open data sharing remains the exception in contemporary systems neuroscience. Here, we take stock of the Allen Brain Observatory, an effort to share data and metadata associated with surveys of neuronal activity in the visual system of laboratory mice. Data from these surveys have been used to produce new discoveries, to validate computational algorithms, and as a benchmark for comparison with other data, resulting in over 100 publications and preprints to date. We distill some of the lessons learned about open surveys and data reuse, including remaining barriers to data sharing and what might be done to address these.},
+	urldate = {2025-10-06},
+	journal = {eLife},
+	author = {de Vries, Saskia EJ and Siegle, Joshua H and Koch, Christof},
+	editor = {Meister, Markus and Gold, Joshua I and Meister, Markus and Chen, Jerry L},
+	month = jul,
+	year = {2023},
+	note = {Publisher: eLife Sciences Publications, Ltd},
+	keywords = {data sharing, electrophysiology, neurophysiology, open science, two photon calcium imaging},
+	pages = {e85550},
+	file = {Full Text PDF:/Users/ashley/Zotero/storage/7FAD93NT/de Vries et al. - 2023 - Sharing neurophysiology data from the Allen Brain .pdf:application/pdf},
+}
+
+@article{de_vries_large-scale_2020,
+	title = {A large-scale standardized physiological survey reveals functional organization of the mouse visual cortex},
+	volume = {23},
+	copyright = {2019 The Author(s), under exclusive licence to Springer Nature America, Inc.},
+	issn = {1546-1726},
+	url = {https://www.nature.com/articles/s41593-019-0550-9},
+	doi = {10.1038/s41593-019-0550-9},
[…]
