### 2025-10-06 6c818c8
Add Python 3.11 environment setup for AllenSDK compatibility
- Created separate requirements files for different lesson groups:
  - requirements_lessons_1-2.txt (Python 3.9+)
  - requirements_lessons_3-4.txt (Python 3.11 required)
  - requirements_greatest_hits.txt (Python 3.9+)
- Updated Setup.ipynb with DandiHub setup instructions for Lessons 3 & 4
- Documented Python 3.11 requirement due to allensdk numpy<1.24 dependency
- Added step-by-step conda environment setup for DandiHub users
