### 2025-10-06 34ed9b7
Document Python 3.11 requirement for building the book
- Updated requirements-docs.txt with Python 3.11 requirement and setup instructions
- Updated CONTRIBUTING.md to specify Python 3.11 environment setup
- Updated README.md Building the Book section with Python 3.11 instructions
- Explains AllenSDK numpy<1.24 compatibility issue with Python 3.12+
