### 2025-10-06 nwb4edu/nwb4edu.github.io#6 @ajuavinett
Hooooo boy this was a doozy but we got there. There are now separate requirement files for each chunk of lessons. There are also new instructions for configuring the Dandihub with a virtual Python 3.11 environment if needed for the allensdk because this is the latest that will work with it and its numpy requirements. The build instructions also contain these details (the entire book only builds with Python 3.11).
