### 2025-10-06 bff8ea1
Exclude Lesson 4-01 from execution during builds
- Added to execution exclude_patterns in _config.yml
- Notebook will use pre-executed outputs instead of re-running
- Prevents empty experiments list issue during builds
- Users can still execute normally on DandiHub
