### 2025-10-07 40b01c1
Exclude Lesson 5-02 from execution during builds
- Added to execution exclude_patterns
- Prevents build errors with jPCA installation
- Users can still execute normally on DandiHub/Jupyter
