### 2025-10-07 8408afd
Clean up repository and fix build configuration
- Update .gitignore to exclude build directories, cache files, and data
- Remove _build directory from version control (generated content)
- Remove Lesson_5/jPCA and nwb-cache directories from tracking
- Fix exclude pattern typo in _config.yml (Analyzing_Two_Photon_Data)
- Remove specific allensdk version check in Lesson_4/03
- Add executed outputs to Lesson_5/02-RotationalDynamics (jPCA plots)
- Update Setup.ipynb with DandiHub Python 3.11 instructions

🤖 Generated with [Claude Code](https://claude.com/claude-code)

Co-Authored-By: Claude <norep
[…]
