### 2025-10-07 0519fed
Remove %matplotlib inline from all notebooks
- Removed %matplotlib inline from 5 notebooks
- This command is not needed in modern Jupyter environments
- Cleaned up: Lesson_4/03-Correlations_in_Two_Photon_Data.ipynb
- Cleaned up: Lesson_3/01-Obtaining_Data_For_Single_Cells.ipynb
- Cleaned up: Data_Science_In_Python/Matplotlib.ipynb
- Cleaned up: Lesson_4/Excitatory_Inhibitory_Decisions.ipynb
- Cleaned up: Lesson_3/Lesson_3_Solutions.ipynb

🤖 Generated with [Claude Code](https://claude.com/claude-code)

Co-Authored-By: Claude <noreply@anthropic.com>
