+++ Lesson_1/000006/dandiset.yaml
@@ -1,0 +1,88 @@
+# DO NOT EDIT THIS FILE LOCALLY. ALL LOCAL UPDATES WILL BE LOST.
+# It can be edited online at https://dandiarchive.org/dandiset/000006
+# and obtained from the dandiarchive.
+'@context': https://raw.githubusercontent.com/dandi/schema/master/releases/0.6.0/context.json
+about: []
+access:
+- contactPoint:
+    schemaKey: ContactPoint
+  schemaKey: AccessRequirements
+  status: dandi:OpenAccess
+assetsSummary:
+  approach:
+  - name: electrophysiological approach
+    schemaKey: ApproachType
+  - name: behavioral approach
+    schemaKey: ApproachType
+  dataStandard:
+  - identifier: RRID:SCR_015242
+    name: Neurodata Without Borders (NWB)
+    schemaKey: StandardsType
+  measurementTechnique:
+  - name: behavioral technique
+    schemaKey: MeasurementTechniqueType
+  - name: spike sorting technique
+    schemaKey: MeasurementTechniqueType
+  - name: surgical technique
+    schemaKey: MeasurementTechniqueType
+  numberOfBytes: 139600500
+  numberOfFiles: 53
+  numberOfSubjects: 12
+  schemaKey: AssetsSummary
+  species:
+  - identifier: http://purl.obolibrary.org/obo/NCBITaxon_10090
+    name: House mouse
+    schemaKey: SpeciesType
+  variableMeasured:
+  - Units
+  - ElectrodeGroup
+  - BehavioralEvents
+citation: Economo, Michael N.; Svoboda, Karel (2024) Mouse anterior lateral motor
+  cortex (ALM) in delay response task (Version draft) [Data set]. DANDI archive. https://dandiarchive.org/dandiset/000006/draft
+contributor:
+- affiliation: []
+  includeInCitation: true
+  name: Economo, Michael N.
+  roleName:
+  - dcite:Author
+  schemaKey: Person
+- affiliation: []
+  identifier: 0000-0002-6670-7362
+  includeInCitation: true
+  name: Svoboda, Karel
+  roleName:
+  - dcite:Author
+  - dcite:ContactPerson
+  schemaKey: Person
+dateCreated: '2020-03-17T13:30:38.667000+00:00'
+description: Extracellular electrophysiology recordings performed on mouse anterior
+  lateral motor cortex (ALM) in delay response task. Neural activity from two neuron
+  populations, pyramidal track upper and lower, were characterized, in relation to
+  movement execution. Some files, as originally (re)distributed from e.g. http://datasets.datalad.org/?dir=/labs/svoboda/Economo_2018
+  were found to be broken and would not be able available among reorganized files
+  under sub-* directories.
+ethicsApproval: []
+id: DANDI:000006/draft
+identifier: DANDI:000006
+keywords: []
+license:
+- spdx:CC-BY-4.0
+manifestLocation:
+- https://api.dandiarchive.org/api/dandisets/000006/versions/draft/assets/
+name: Mouse anterior lateral motor cortex (ALM) in delay response task
+protocol: []
+relatedResource:
+- identifier: doi:10.1038/s41586-018-0642-9
+  name: Distinct descending motor cortex pathways and their roles in movement
+  relation: dcite:IsDescribedBy
+  repository: Nature
+  resourceType: dcite:JournalArticle
+  schemaKey: Resource
+  url: https://www.nature.com/articles/s41586-018-0642-9
+repository: https://dandiarchive.org
+schemaKey: Dandiset
+schemaVersion: 0.6.0
+studyTarget: []
+url: https://dandiarchive.org/dandiset/000006/draft
+version: draft
+wasGeneratedBy: []
