### 2025-10-07 b7b4b4e
Exclude PlaceFields notebook and add note about interactive plots
- Added Lesson_5/01-PlaceFields.ipynb to exclude_patterns in _config.yml
- Added note explaining that Plotly interactive visualizations won't show on static website
- Directed users to run notebook on DandiHub to see grid cell visualizations
- Linked to Using this Book page for DandiHub instructions

🤖 Generated with [Claude Code](https://claude.com/claude-code)

Co-Authored-By: Claude <noreply@anthropic.com>
