### 2025-10-07 37ef4c5
Update paper.bib
