+++ paper.bib
@@ -38,7 +38,7 @@
 }
 
 @software{executable_books_project_2021_4539666,
-  author       = {Executable Books Project},
+  author       = {Executable Books Project,},
   title        = {Jupyter Book},
   month        = feb,
   year         = 2021,
