### 2025-10-07 nwb4edu/nwb4edu.github.io#6 @arokem
I think this is fine for the purpose of the JOSE paper, but if you want these materials to have longevity you might want to alert the Allen SDK developers (if such are still around) that [SPEC0](https://scientific-python.org/specs/spec-0000/) recommends dropping support for 3.11 pretty soon, so it will soon be difficult to use the Allen SDK with the rest of the scientific Python ecosystem. Again, doesn't seem like a deal-breaker for the book, but worth thinking about more generally.
