### 2025-10-07 468f256
Move PNG files to _images folder and update config
