+++ _config.yml
@@ -1,5 +1,5 @@
 title: Teaching & Learning with NWB Datasets
-logo: nwb_learning.png
+logo: _images/nwb_learning.png
 only_build_toc_files: true
 repository:
   url: https://github.com/nwb4edu/nwb4edu.github.io
