### 2025-10-07 b4666a0
Add Lesson 3 notebooks to exclude patterns to prevent allensdk install errors
