### 2025-10-07 3572590
Add version 1.0.0 for JOSE release
