+++ _config.yml
@@ -1,4 +1,5 @@
 title: Teaching & Learning with NWB Datasets
+version: 1.0.0
 logo: _images/nwb_learning.png
 only_build_toc_files: true
 repository:
