### 2025-10-07 cdf1aec
Update copyright and author information
