### 2025-10-07 80dd8a0
Update CONTRIBUTING.md
remove unnecessary details about markdown formatting & simplify
