### 2025-10-07 #309 @arokem
Should we wait for [this PR](https://github.com/nwb4edu/nwb4edu.github.io/pull/1) to be merged before re-reviewing? Looks like it was opened a while ago, but there are also very recent commits on it. Or in other words, could you please let us know here when you would like us to take another look?
