### 2025-10-07 01bbdf7
Remove CODE_ISSUES_TO_FIX.md task list
🤖 Generated with [Claude Code](https://claude.com/claude-code)

Co-Authored-By: Claude <noreply@anthropic.com>
