### 2025-10-10 #309 @arokem
Just to be clear: are you ready for the next round of review? Do you want us to take another look now?
