### 2025-10-14 #309 @ajuavinett
Yes, ready for the next round of review!

(Sorry, I thought this would be @JasonJWilliamsNY 's call.)
