### 2025-10-20 #309 @TomDonoghue
I've had a check through the responses to the issues and updates to the resource, and I think the updates address all of my comments and suggestions! I've revisited the review checklist and now checked everything off as I think everything meets the spec for the JOSE review criteria. Let me know if there's anything else needed from me - but as is, I'm happy to sign off on this as ready for publication!

Congrats / thank you @ajuavinett for developing a very useful resource and making this review process very smooth!
