### 2025-10-20 #309 @arokem
I have now also taken another look and checked off all the remaining boxes on my list. Great work on this revision! Also happy to sign off on this as ready for publication.
