### 2025-11-05 f285f62
suggested grammar clarifications
