### 2025-11-10 #309 @JasonJWilliamsNY
Thanks @ajuavinett one more thing...  Can you check off the items in the comment above "Additional Author Tasks After Review is Complete" — just want to make sure everything is correct.

Double check authors and affiliations (including ORCIDs)
Make a release of the software with the latest changes from the review and post the version number here. This is the version that will be used in the JOSE paper.
Archive the release on Zenodo/figshare/etc and post the DOI here.
Make sure that the title and author list (including ORCIDs) in the archive match those in the JOSE paper.
Make sure that the license listed for the archive is the same as the software license.
