### 2025-11-11 #309 @ajuavinett
Sure thing, just made a few final tweaks to author info and the release is now archived in Zenodo:

✅ Double check authors and affiliations (including ORCIDs)
✅ Make a release of the software with the latest changes from the review and post the version number here: **v1.1.0**
✅ Archive the release on Zenodo/figshare/etc and post the DOI here: [https://zenodo.org/records/17575041 ](https://zenodo.org/records/17575041 )
✅ Make sure that the title and author list (including ORCIDs) in the archive match those in the JOSE paper.
✅ Make sure that the license listed for the archive is the same as the software license.
