### 2025-11-12 #309 @JasonJWilliamsNY
Set version with @editorialbot set 1.1.0 as version
