### 2025-11-12 #309 @editorialbot
:wave: @openjournals/jose-eics, this paper is ready to be accepted and published.
