### 2025-12-10 #309 @labarba
Congratulations @ajuavinett – your JOSE paper is published! 🚀

Huge thanks to our Editor: @JasonJWilliamsNY and Reviewers: @TomDonoghue, @arokem — your contributions make this possible 🙏
