### 2026-05-02 #309 @ajuavinett
Hi @labarba & @JasonJWilliamsNY, I'm not sure exactly how this happened (whether it was a copy paste error on my part or AI-confabulation in using Claude to prepare the release) but unfortunately the ORCID linked to the paper is incorrect. It should be [0000-0002-4254-3009](https://orcid.org/0000-0002-4254-3009), which is now reflected in https://github.com/nwb4edu/nwb4edu.github.io/blob/master/paper.md.

I'm very sorry for the extra work, but if it's not too much, could this be corrected?
